(proxy): Move necessary functions from tor
